refactor: 결제 카드 관련 UI 컴포넌트 이름 변경 및 패키지 이동
