+++ app/src/main/java/woowacourse/payments/ui/cards/SingleCardsSection.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import woowacourse.payments.ui.components.PaymentCreateCard
+import woowacourse.payments.ui.newcard.NewCard
 import woowacourse.payments.ui.model.PaymentCardUiModel
 import woowacourse.payments.ui.preview.paymentCardUiModelSample
 
@@ -26,7 +26,7 @@ fun SingleCardsSection(
     ) {
         Spacer(Modifier.height(12.dp))
         PaymentCard(card)
-        PaymentCreateCard(onAddClick)
+        NewCard(onAddClick)
     }
 }
